requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
iniconfig==1.1.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytest-env==0.6.2
pytest-order==0.9.4
python-box==5.2.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
